java programmimg question added

+++ src/javainterviewprogram/ReverseAnInteger.java
@@ -3,16 +3,18 @@ package javainterviewprogram;
 public class ReverseAnInteger {
 
 	public static void main(String[] args) {
-   int num=87654;
-   int rev=0;
+   long num=87654;
+   long rev=0;
    while(num!=0){
-	   rev= rev*10+num%10;
-	   num=num/10;
+	   rev= rev*10 + num%10;
+	   num= num/10;
 	   
    }
         System.out.println("reverse num is :"+rev);
+        
+        
       //2-Using StringBuffer method
-        int num1=12345;
+       long num1=12345;
         System.out.println(new StringBuffer(String.valueOf(num1)).reverse());
 	}
